created TitleFilter.py removed all functions from MTG_title_reader.py to TitleFilter.py

+++ MTG_title_reader.py
@@ -7,10 +7,8 @@ import cv2
 from os import listdir
 import pytesseract
 
-from pytesseract import Output
-
 from CardImage import CardImage
-from DebugImage import DebugImage
+from TitleFilter import TitleFilter
 
 
 def url_to_image(url, readFlag=cv2.IMREAD_GRAYSCALE):
@@ -22,31 +20,6 @@ def url_to_image(url, readFlag=cv2.IMREAD_GRAYSCALE):
 
     # return the image
     return image
-
-
-# <editor-fold desc="filter title">
-def filter_title(texts):
-    original_title = ''
-    for i in range(len(texts)):
-        if texts[i] != "" and texts[i] != " ":
-            if not any(ext in texts[i] for ext in testStr):
-                if original_title != '':
-                    original_title += ' '
-                original_title += texts[i]
-    return original_title
-
-
-def remove_word_from_title(full_title, removeFirstWord=True):
-    split_title = full_title.split()
-    for word in range(len(split_title)):
-        if removeFirstWord is True and word == 0:
-            split_title.pop(word)
-        if removeFirstWord is False and word == len(split_title) - 1:
-            split_title.pop(word)
-    return ' '.join(map(str, split_title))
-
-
-# </editor-fold>
 
 
 # noinspection PyShadowingNames
@@ -92,7 +65,7 @@ def get_exact_card(card_list, card_img):
         return card_list[index_highest_card]
 
 
-def card_searching_single(txt_title, card_img):
+def card_searching_single(txt_title, card_img, filter_text):
     card_list = []
     if txt_title != "":
         cards = get_cards(txt_title)
@@ -101,7 +74,7 @@ def card_searching_single(txt_title, card_img):
         if len(cards) == 0:
             print(txt_title + " not found in the db"
                               "\n try to remove last word ... ")
-            title_r = remove_word_from_title(txt_title, False)
+            title_r = filter_text.remove_word_from_title(txt_title, False)
             cards = get_cards(title_r)
 
             for card_index in range(len(cards)):
@@ -111,7 +84,7 @@ def card_searching_single(txt_title, card_img):
             if len(cards) == 0:
                 print(title_r + " not found in the db"
                                 "\n try to remove First word ... ")
-                title_r = remove_word_from_title(txt_title)
+                title_r = filter_text.remove_word_from_title(txt_title)
                 cards = get_cards(title_r)
 
                 for card_index in range(len(cards)):
@@ -133,13 +106,13 @@ def card_searching_single(txt_title, card_img):
 
 
 # constants
-testStr = [":", "’", ";", "—", "$", "/", "_", "|"]
 os.putenv('TESSDATA_PREFIX', 'C:\\Program Files\\Tesseract-OCR\\tessdata\\mtg.traineddata')
 pytesseract.pytesseract.tesseract_cmd = 'C:\\Program Files\\Tesseract-OCR\\tesseract.exe'
 
 # variables
 y1, y2 = 15, 90
 x1, x2 = 30, 410
+testStr = [":", "’", ";", "—", "$", "/", "_", "|"]
 
 amount_of_cards_not_found = 0
 
@@ -149,7 +122,7 @@ cards_not_found = back_mtg_card
 
 is_debugging = False
 card_image = CardImage(is_debugging)
-
+text_filter = TitleFilter(is_debugging)
 for f in listdir("TestCards"):
     print(f)
     filename = "TestCards/" + f
@@ -157,15 +130,10 @@ for f in listdir("TestCards"):
     # get fixed img from large image
     img = card_image.filter_card_from_img(img, f)
     # get the title from the image
-    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-    title = gray[y1:y2, x1:x2]
-    flag, thresh = cv2.threshold(title, 100, 255, cv2.THRESH_BINARY)
-    if is_debugging is True:
-        DebugImage.write_image(thresh, '6_title_thresh', f)
-        DebugImage.write_image(title, '7_title_img', f)
-    title_txt = filter_title(pytesseract.image_to_data(title, output_type=Output.DICT)['text'])
-    print(title_txt)
-    card = card_searching_single(title_txt, img)
+
+    title_txt = text_filter.get_title_from_image(img, f)
+
+    card = card_searching_single(title_txt, img, text_filter)
     if card is None:
         img_resized = cv2.resize(img, size_card)
         cards_not_found = np.concatenate((cards_not_found, img_resized), axis=0)
@@ -175,9 +143,10 @@ amount_of_cards = len(os.listdir("TestCards"))
 print(f'{amount_of_cards_not_found}/{amount_of_cards} not found')
 cv2.imshow('cards not found', cards_not_found)
 cv2.waitKey(0)
-val = input("Delete debug images? y/n: ")
-if val == 'y' or val == "Y":
-    shutil.rmtree('debug')
-    print("debug images deleted...")
-else:
-    pass
+if is_debugging is True:
+    val = input("Delete debug images? y/n: ")
+    if val == 'y' or val == "Y":
+        shutil.rmtree('debug')
+        print("debug images deleted...")
+    else:
+        pass
